Did a bunch of weird testing chnages. Good luck changing em, vijay

+++ control/kalman/kalman.py
@@ -35,7 +35,7 @@ class KalmanFilter(object):
         return self.current_state_estimate
         
     def getCurrentAngle(self):
-        return int(self.current_state_estimate[0])
+        return float(self.current_state_estimate[0])
 
     def Step(self, control_vector, measurement_vector):
         now = time.time()
